move batched multirotor to multirotor.py
