+++ tests/test_batched_sims.py
@@ -2,8 +2,7 @@ import numpy as np
 import torch
 
 from rotorpy.trajectories.minsnap import BatchedMinSnap
-from rotorpy.vehicles.batched_multirotor import BatchedMultirotor, BatchedDynamicsParams
-from rotorpy.vehicles.multirotor import Multirotor
+from rotorpy.vehicles.multirotor import Multirotor, BatchedDynamicsParams, BatchedMultirotor
 from rotorpy.controllers.quadrotor_control import BatchedSE3Control, SE3Control
 from rotorpy.vehicles.crazyflie_params import quad_params as cf_quad_params
 from rotorpy.vehicles.hummingbird_params import quad_params as hb_quad_params
